Refactoring

+++ src/main/java/ludo/mentis/aciem/scl/util/WebUtils.java
@@ -90,7 +90,7 @@ public class WebUtils {
         next.setUrl(getStepUrl(page, page.nextOrLastPageable().getPageNumber()));
         steps.add(next);
 
-        final long rangeStart = page.getNumber() * page.getSize() + 1l;
+        final long rangeStart = (long) page.getNumber() * page.getSize() + 1L;
         final long rangeEnd = Math.min(rangeStart + page.getSize() - 1, page.getTotalElements());
         final String range = rangeStart == rangeEnd ? "" + rangeStart : rangeStart + " - " + rangeEnd;
         final PaginationModel paginationModel = new PaginationModel();
